utils release and user
