+++ utils.go
@@ -1,0 +1,46 @@
+package main
+
+import (
+	"bytes"
+	"compress/gzip"
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+)
+
+func decodeGzip(bodyBytesInput []byte) (bodyBytesOutput []byte) {
+	bodyReader, err := gzip.NewReader(bytes.NewReader(bodyBytesInput))
+	if err != nil {
+		fmt.Println(err)
+	}
+	bodyBytesOutput, err = ioutil.ReadAll(bodyReader)
+	if err != nil {
+		fmt.Println(err)
+	}
+	return
+}
+
+func encodeGzip(b []byte) bytes.Buffer {
+	var buf bytes.Buffer
+	w := gzip.NewWriter(&buf)
+	w.Write(b)
+	w.Close()
+	// return buf.Bytes()
+	return buf
+}
+
+func unmarshalJSON(bytes []byte) map[string]interface{} {
+	var _interface map[string]interface{}
+	if err := json.Unmarshal(bytes, &_interface); err != nil {
+		panic(err)
+	}
+	return _interface
+}
+
+func marshalJSON(body map[string]interface{}) []byte {
+	bodyBytes, errBodyBytes := json.Marshal(body)
+	if errBodyBytes != nil {
+		fmt.Println(errBodyBytes)
+	}
+	return bodyBytes
+}
